\name{unuran.discr-class}
\docType{class}

\alias{unuran.discr-class}
\alias{print,unuran.discr-method}
\alias{show,unuran.discr-method}
\alias{initialize,unuran.discr-method}

\title{Experimental class "unuran.discr"}

\description{%
  The class "unuran.discr" provides am interface to UNU.RAN objects for
  discrete distributions.
  \strong{Experimental. Do not use!}
}

\author{
  Josef Leydold \email{leydold@statmath.wu-wien.ac.at} and
  Wolfgang H\"ormann \email{hormannw@boun.edu.tr}
}

\references{
  J. Leydold and W. H\"ormann (2000-2007):
  UNU.RAN User Manual, see
  \url{http://statmath.wu-wien.ac.at/unuran/}.
}

\seealso{
  \code{\link{unuran-class}}.
}

\keyword{classes}
\keyword{distribution}
\keyword{datagen}
