\name{unuran.distr-class}
\docType{class}

\alias{unuran.distr-class}
\alias{print,unuran.distr-method}
\alias{show,unuran.distr-method}
\alias{initialize,unuran.distr-method}

\title{Experimental virtual class "unuran.distr"}

\description{%
  The virtual class "unuran.distr" provides am interface to UNU.RAN
  objects for distributions.
  \strong{Experimental. Do not use!}
}

\author{
  Josef Leydold and Wolfgang H\"ormann
  \email{unuran@statmath.wu-wien.ac.at}.
}

\references{
  J. Leydold and W. H\"ormann (2000-2007):
  UNU.RAN User Manual, see
  \url{http://statmath.wu-wien.ac.at/unuran/}.
}

\seealso{
  \code{\link{unuran-class}}.
}

\keyword{classes}
\keyword{distribution}
\keyword{datagen}
