\name{unuran.details}
\docType{function}

\alias{unuran.details}

\title{Information on a given "unuran" generator object}

\description{
  Prints type of \code{unuran} generator, data used from distribution, 
  parameter for algorithm, performance characteristic, and hints to
  adjust the performance of the generator.
  
  [Advanced] -- Print object.
}

\usage{
unuran.details(unr)
}

\arguments{
  \item{unr}{an \code{unuran} object.}
}

%% \details{ }

\seealso{%
  \code{\linkS4class{unuran}}.
}

%%\note{ }

\author{
  Josef Leydold and Wolfgang H\"ormann
  \email{unuran@statmath.wu.ac.at}.
}

%%\examples{ }

\keyword{distribution}
\keyword{datagen}

